Bib record browser with 'see also', etc from linked authority headings
This feature provides a patron-oriented OPAC interface for browsing
bibliographic records.

Users choose to browse by Author, Title, Subject, or Series. They then
enter a browse term, and the nearest match from a left-anchored search
on the headings extracted for browse purposes will be displayed in a
typical backwards/forwards paging display. Headings link to search
results pages showing the related records. If the browse heading is
linked to any authority records, and if any *other* authority records
point to those with "See also" or other non-main entry headings, those
alternative headings are displayed a linked to a search results page
showing related bib records related to the alternate heading.

The counts of holdings displayed next to headings from bibliographic
records are subject to the same visiibility tests as search. This means
that the org unit (and copy location group) dropdown on the browse
interface affects counds, and it further means that whether or not
you're looking at the browse interface through the staff client makes a
difference.

This builds on the two previous commits that provide inter-authority
linking and the linking of metabib.browse_entry rows to authority
records.

This also contains, in squashed form, these commits that resulted from
collaboration on LP #1177810:

    Two bugfixes to OPAC Browse: non-filing indicators, leading-article warning
    Fix paste-o encountered in Bib browse upgrade script
    OPAC Browse: fix 0-9 link in paging shortcuts; padding issues
    OPAC Browse: Improve authority code to show more headings
    OPAC Browse: Fix authority counting
    Extensions to our MODS32 stylesheet to capture all possible NFI in fields
    OPAC Browse: i18n improvement for short terms
    OPAC Browser: Display Public General Notes from authorities when possible
    OPAC Browse: Build browse entry sort_value column separately from value
    OPAC Browse: We don't want role/relator info in browse headings
    OPAC Browse: Better display of tracings from authorities
    OPAC Browse: Pick up authority links from 650 fields
    OPAC Browse: Show authority tracings only when inter-authority links exist
    OPAC Browse: use  superpage concept for performance; fix other counting bug
    OPAC Browse: Fix broken authority reference link

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>
Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>