LP 1169193 transit_slip macro for copy status
Adds support for a new 'transit_slip' receipt template macro called
"transit_copy_status", which displays the stored status of the
in-transit copy (i.e. the status the copy will be placed in once it
arrives at the transit destination).  This is useful, in particular, for
lost or long-overdue items which are not "immediately available" upon
checkin.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>