LP 1169193 Bypass longoverdue w/ default A/T runner
Avoid processing LONGOVERDUE circulations in the default
action_trigger_runner.pl filter.

When no custom filters are used, the script will provide it's own
default filter for circulation-based hooks.  This change ensure that
when the default filter is in use, the script will not attempt to
process longoverdue circulations, since they have already been processed
(which is how they became longoverdue in the first place).

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>