LP1171875 Add locale support to IDL2js
/IDL2js now reads locale information from either locale= CGI parameter
or Accept-Language HTTP headers.  The locale-aware IDL is loaded from
/reports/fm_IDL.xml via Apache subrequest.  Each full copy of the IDL is
cached within the Apache processes to avoid the need to re-parse the IDL ad
infinitum for full IDL retrieval.  Partial IDL retrieval is also supported
(but not cached).

No attempt is made to cleanse the locale -- invalid locale strings are
discarded -- so it's the callers responsibility to pass a valid locale.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Pasi Kallinen <pasi.kallinen@pttk.fi>
Signed-off-by: Mike Rylander <mrylander@gmail.com>