Shutdown added content sockets to prevent hung connections.
We were getting several thousand sockets hanging around in a
CLOSE_WAIT state on our server.  The connections were from our
server and to our server.  Further investigation showed it to be
caused by the stage2 added content lookup not closing the request
sockets when it was done reading data.  This branch fixes that.

Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>
Signed-off-by: Mike Rylander <mrylander@gmail.com>