"Return to record" after submission of phone number on SMS page fixed
Before this, it only worked on the page load *before* form submission.
There's probably a more elegant solution that doesn't involve keeping
track of URIs, but this seems to me to work.

Problem relayed to me by Dale Rigney at Equinox, who helped reproduce
the issue and test a fix.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Jason Etheridge <jason@esilibrary.com>