When delivering MARC21 via SuperCat, OpenSearch, offer filename w/ extension
If you want a record feed in MARC21, without this commit a web browser
just suggests "marc21" as the filename for saving to your local system.

Although many consumers of these feeds will be custom scripts and other
non-interactive software components that won't care about filename, for
the web browser case it's helpful if the server suggests a filename that
ends in .mrc.

So here we go.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Ben Shum <bshum@biblio.org>