TPAC - Password Messages
In jspac a message appeared when the patron went to change their
password stating the requirements. This same message is missing
in TPAC. The wording from jspac has been added to a separate
file, update_password_msg.tt2,  that can later be easily edited
and the original update_password.tt2 has been edited with an
INCLUDE statement pointing to the newly created file.

Also improved is the message when requesting a forgotten password.
Patrons are unaware that a valid email address needs to be in
their record until after they try to send a reset. Text has been
added to a new file, password_reset_msg.tt2,  to explain the need
for an email address.  Password_reset.tt2  has been edited with
an INCLUDE statement pointing to the newly created file.

Style.css.tt2 was also updated to remove extra padding on
#account-update-email table to raise the table on
update_password.tt2.  This value is also used on
update_username.tt2 and update_email.tt2 without considerable
impact to the look of those pages.

Signed-off-by: Melissa Lefebvre <mlefebvre@biblio.org>
Signed-off-by: Dan Scott <dscott@laurentian.ca>