Fix LP1180140, View Holds not working for a serial with subscription and no issuances
A request is made to
open-ils.circ.holds.retrieve_all_from_title,
which is the function Circ/Holds.pm/all_rec_holds().

In the function, when checking for issuance holds, there is a missing @
to dereference the list of issuance IDs, so that the ahr table is search
with badly formed input, which results in returning an undefined list
instead of an empty list.  The error will occur for any record with a
serial.subscription entry but with no related entries in
serial.issuance.

We also recode the initialization of the response object in order to
clarify its expected structure, a hash of array references.

Signed-off-by: Steven Chan <schan@sitka.bclibraries.ca>
Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>