Expose conjoined bibs in TPAC
This initial work is based on the GRPL need to handle eReader bib/copy display.
The "joined" titles will display a copy record with a call_number link back to
the "master" bib, where the master will display a set of links to individual
titles from each "copy".

TODO: consider peer_type when determining whether or not to display

Signed-off-by: Bill Ott <bott@grpl.org>
Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Ben Shum <bshum@biblio.org>