Fix Launchpad Bug 1183418.
Clicking on an author with a dash or dates in the name on the search
results page comes up with 0 results. It looks like the dash is stripped
from the date and/or name.

Clicking on such a link from the record view actually works because
the dash is replaced with a space.

This branch changes table.tt2 to replace special characters in author
output with a space in the same way that authors.tt2 does.

Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>
Signed-off-by: Ben Shum <bshum@biblio.org>