Patron Editor can enter erroneous values for Claims-returned count
In the staff client, Patrons interface, Edit screen, when
mouse-scrolling up and down the Edit form, it is possible to
inadvertently change the value in Claims-returned Count or Claims Never
Checked Out Count field.  It will happen if the mouse hovers over either
data field while using the mouse wheel.

Both data fields use the dijit.form.NumberSpinner widget to provide a
'spin' behaviour.  The fix involves cancelling the mouse scroll event as
it propagates to input fields in table rows that are using the
dijit.form.NumberSpinner widget, because the mouse scroll would be
applied too early by the widget, before the user has intentionally
focussed on one of the input fields.  Now, mouse scrolling has no effect
on the two input fields; it will only have an effect in scrolling the
page up or down.

Signed-off-by: Steven Chan <schan@sitka.bclibraries.ca>
Signed-off-by: Ben Shum <bshum@biblio.org>