LP1185524 - Duplicate patron checking in the user editor is limited to workstation OU
Added a new library setting opac.duplicate_patron_check_use_consortium.
When true, the duplicate check will use the consortium (1) as the OU.
When false or not set, the duplicate check will use the workstation OU.

The setting is checked in subroutine that creates the search request and
sets the OU accordingly.

[LFW: Signing off but following with material changes next commit.]

Signed-off-by: Kyle Tomita <ktomita@catalystitservices.com>
Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>
Signed-off-by: Chris Sharp <csharp@georgialibraries.org>