Bring back "split-brain" indexing normalization
Before 2.4-era changes to indexing and search, normalizers were applied
in a way that resulted in different effects on the value and index_vector
columns of metabib.*field_entry tables.  This behavior was lost during
the changes mentioned above, but we need it for several things to work
properly, phrase search and (likely) facets among them.  This commit
brings that behavior back.

See here for the original behavioral documentation:

http://evergreen-ils.org/dokuwiki/doku.php?id=documentation:indexing#field_normalization_settings

Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Dan Scott <dscott@laurentian.ca>