Support script-based circ in nearest_hold()
The nearest_hold() implementation recently changed to pull in the circ
library ID; however, if circulation runs through scripts, the circ
library on the copy object has been fleshed, and the nearest hold
routine gets very confused.

Check to see if we do, in fact, have an object instead of a plain
scalar, and assign the value to a placeholder variable accordingly.

Thanks to Jeff Godin for his great assistance on debugging this!

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Mike Rylander <mrylander@gmail.com>