Only attempt to map copies once per hold
The development of Calculated Proximity Adjustment (CPA) for 2.4 added a "prox"
column to action.hold_copy_map (ahcm) to store on variant of copy proximity.  That
value is calculated at the same time that certain other data structures are
set up, and is conceptually tied to them, so the resulting combined data
set is used to insert values into the ahcm.  However, user configuration
can cause that data set to contain non-unique values, where the data structure
used previous to CPA was guaranteed to be unique.  This lack of uniqueness
would then violate a constraint on the achm table.  This commit restores the
required uniqueness for the purpose of creating the ahcm rows.

Signed-off-by: Mike Rylander <mrylander@gmail.com>
Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>

Conflicts (whitespace only):
	Open-ILS/src/perlmods/lib/OpenILS/Application/Storage/Publisher/action.pm