LP#1308239 Support pre-cat copy hold targeting
Remove an unused line of code in the hold targeter which was preventing
targeting of pre-cat copies.

In OpenILS::Utils::PermitHold::permit_copy_hold, neither in-db nor
script based hold viability checks inspect the value of the
"title_descriptor" field in the provided params.  Remove this field and
pre-cats just work.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Ben Shum <bshum@biblio.org>