Staff client index: Add tooltips, restore spacing
This commit adds to Pasi's improvements of the staff client index (or
portal) page by adding a tooltip message to all of the "open in a new
tab" icons, using the already translatable alt text of the icon. Also,
this commit restores the original padding on the rows, which was changed
when border-collapse was previously added.

Signed-off-by: Remington Steed <rjs7@calvin.edu>
Signed-off-by: Dan Wells <dbw2@calvin.edu>