Maintenance message: avoid Unicode corruption, support KPAC
By shifting from an Apache config environment variable to a config.tt2
variable, we gain the advantage of more robust Unicode support as well
as the ability to turn maintenance messages on or off without requiring
a reload of Apache.

However, we do lose the convenience of per-vhost maintenance messages;
sites requiring this functionality can use the ctx.hostname variable to
provide conditional per-host messages.

This also adds support for displaying the maintenance message in the
KPAC.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Ben Shum <bshum@biblio.org>