Conjoined items / foreign items display in TPAC
Foreign items now appear at the top of the copy table, no matter what
copy page you are on. See "The two violin concertos" in the Concerto
data set for an example; the foreign bib is linked, with a label for the
designated peer type.

The parent bib record displays all of the child records in the copy table
underneath the pertinent copy; each title / author entry links to the
child bib display, naturally. See "Violin concerto no. 2 ; Six duos
(from 44 Duos)" in the Concerto data set for an example.

We escape all of the text retrieved from the database to avoid XSS
(thanks for the heads-up from Pasi Kallinen).

Signed-off-by: Dan Scott <dscott@laurentian.ca>

Now with HTML escaping for safety

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Ben Shum <bshum@biblio.org>