LP#1202742 Non-active transit copy status messages
When an in-transit copy is checked into the staff client, display a special
message in the transit alert dialog and in the printed transit receipt
(optionally, via macro) if the copy is in (or, rather, will be once it
arrives at its destination) a non-active copy status.

See config.copy_status.copy_active.

For example, assuming the org unit setting 'circ.lost_immediately_available'
is unset, when a Lost copy is checked in that must transit home, the
following message will appear in the transit alert dialog:

This item is in status "Lost", additional staff action may be required.

Additionally, the value of the 'transit_copy_status_msg' macro, which
defaults to "", will be set to this message, so that the message may
appear in printed transit slips.

Note that the code will test for the presence of the
"staff.circ.utils.transit.copy_status_message" string property and fail
gracefully if it is unset.  Through this, admins can disable this
feature entirely.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Chris Sharp <csharp@georgialibraries.org>
Signed-off-by: Kathy Lussier <klussier@masslnc.org>