LP1204273 - Repair org unit settings for making state not required
As noted by tsbere in IRC, the org unit settings were using "au" instead of "aua"
which led to issues applying the settings and making the addresses require the
state field.

Also, change the constraint for field_name on actor.org_unit_setting_type_log to
be "DEFERRABLE INITIALLY DEFERRED" to be more flexible and consistent with other
parts of the database relating to org unit setting names.

And, the table auditor.actor_usr_address_history needs to be updated so that
state can be NULL there as well.

Signed-off-by: Ben Shum <bshum@biblio.org>
Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>