LP#1418772: Avoid internal server error on viewing full record when copy create_date is null
In the TPAC client, when it tries to show record details containing a copy
record with no create date, it shows an Internal Server Error instead.

1. The error is caused by trying to execute the parse_datetime() method
in the parse_date() function in the WWW/EGCatLoader/util.pm module with
an empty date string. The function will normally translate a datetime
string from the database to a datetime string that is formatted for TPAC
templates.  The fix is to not execute parse_datetime() and just return
an empty string.

2. In the record/copy_table.tt2 template, if an empty datetime string is
the value for copy_info.create_date, the format() method of the Date
plugin will show the current datetime by default. The fix is to show '-'
in its place, replicating the same template logic as for
copy_info.due_date.

Signed-off-by: Jeff Davis <jdavis@sitka.bclibraries.ca>

Conflicts:
	Open-ILS/src/templates/opac/parts/record/copy_table.tt2

Signed-off-by: Galen Charlton <gmc@esilibrary.com>