Allow action trigger events to be repeated
Currently, there is no way to repeat an event (action/trigger) when a
real-world event occurs more than once for the same target. This only
matters for targets that are persistent (such as a library card)
combined with events that can reoccur (such as expiration, if the
expiration can be removed). To address this, we add a new field to the
action_trigger.event_definition table that sets the amount of time
after which the event is allowed to repeat.

This commit adds the new field to the IDL, database schema and template
files and also provides the simple db upgrade script. It is worth noting
that we chose to base the repeatability on the start_time of the previous
event, which is more predictable than complete_time (which could be null
if the event fails).

Signed-off-by: Remington Steed <rjs7@calvin.edu>
Signed-off-by: Mike Rylander <mrylander@gmail.com>