LP1207396 TPAC org_selector supports org filter
When building and org unit selector via the build_org_selector MARCO,
the caller can now pass in a "valid_org_list" array which contains org
unit IDs.  When such a list is provided, any org units in the selector
which are not in the list are marke as disabled.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Dan Wells <dbw2@calvin.edu>