LP 1084732 TPAC copy location selector
Copy location filter selector for the TPAC advanced search interface.
The behavior is the same as the analogous JSPAC feature.  The interface
components are hidden by default, so non-JS browsers will be none the
wiser.

A number of small tweaks to related code were needed and are included in
this commit:

1. Make copy locations field_safe in the IDL so they can be retrieved
   and parsed without the IDL
2. Insert a (full) hash-based org unit tree as inline JS at the bottom
   of js.tt2 so JS will have access to the whole tree without having to
   load the IDL to understand it.
3. Fix a bug in openils.CGI : decodeURIComponent parameter names

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Ben Shum <bshum@biblio.org>