Billing UI style lost and longoverdue circulations
When a LOST or LONGOVERDUE circulation, which has not yet been returned,
appears in the patron's list of billed transactions, style them
differently from other not-yet-returned items.

Similar to regular not-yet-returned items, display a helpful message to
staff indicating what colors map to what types of circulations.

The default colors are maroon for LOST and organge for LONGOVERDUE.

To change the colors, a change to circ.css (for the style) and lang.dtd
(for the hint) are required.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Mike Rylander <mrylander@gmail.com>