Schema.org authors: relate properties to the right context
While the approach of using the existing <a href> to launch a new search
based on the author name was workable, it may have been confusing for
schema.org processors such as Google Rich Snippets who expected that URI
to identify a resource for the author, not a set of search results.

Using @resource / @about attribute pairing, and wrapping the whole type
in an extra <span> element, enables us to avoid associating the
properties within the <a href> with the implicitly chained type.

Signed-off-by: Dan Scott <dscott@laurentian.ca>
Signed-off-by: Ben Shum <bshum@biblio.org>