LP#856688 OUS to disable org unit as hold pickup lib
YAOUS that removes an org unit from consideration as a hold pickup
library.  This setting only affects the OPAC display and does not
otherwise affect the display of the org unit in the OPAC.  (In other
words, it's not the same thing as actor.org_unit.opac_visible).  It also
has no affect on hold targeting / capturing.

[LFW: Numbered upgrade script.]

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@esilibrary.com>