Vandelay copy overlay call number merge
When overlaying a copy, if the copy brings with it a new call number
which collides with an existing call number AND said copy is the only
copy linked to its (old) call number, merge the copy's (old) call
number into the (new) existing call number, i.e. the one we collided
with.

Signed-off-by: Bill Erickson <berick@esilibrary.com>
Signed-off-by: Dan Wells <dbw2@calvin.edu>