'Opportunistic' Acq In-process Copy Overlay
It is both normal and common to overlay brief acquisitions bib
records with more complete records, and it is simple to create
copies as part of the overlay process, but there is no *simple* way
to have the imported copies overlay the acq copies.

This code builds off the existing copy overlay code (which matches
and overlays using specified IDs), but uses broader matching criteria.
By selecting the new option, "Auto Overlay In-process Acquisitions
Copies", the system will potentially overlay copies which:
  * have associated lineitem details (that is, they were created
    by the acquisitions process), and
  * that lineitem detail has the same owning_lib as the incoming
    copy's owning_lib
  * the current copy associated with that lineitem detail is "In
    process"

Also, fix two small bugs, one which prevented overlay using the
'Available' status (change 'if $val' to 'if defined $val', since
'Available' is '0'), and another which prevented item overlay when
selecting the match record manually (as_imported was null in the DB).

Signed-off-by: Dan Wells <dbw2@calvin.edu>
Signed-off-by: Mike Rylander <mrylander@gmail.com>