Use xact_start for due date calc instead of now
If you don't supply a due date when creating a circulation with the
open-ils.circ.checkout family of calls, the due date is currently
calculated from now().  This code modification changes that so that
if a checkout_date is supplied that is used to calculate the due date
instead of now().

This enables some interesting uses of the open-ils.circ module for
batch circulation in the cases of offline use from a spreadsheet and
not the offline client or in the case of migrating circulations from
another ILS into Evergreen.

Signed-off-by: Thomas Berezansky <tsbere@mvlc.org>
Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>
Signed-off-by: Dan Wells <dbw2@calvin.edu>