Serials: alert notes when receiving
This commit adds the ability to flag serial notes as receive-time
alerts, and also adds the interface widgets necessary to check for
and see these alerts in the serial control receiving interface.

notes.xul was modified extensively to allow for an arbitrary pile of
notes to be displayed, rather than just displaying the notes for a
single object (which it still can do).

As it stands, the notes are compiled locally from already retrieved
objects, but it would also be reasonable to add a middle-layer
function to do roughly the same thing.  Depending on how this is done
it could require a fair amount of logic duplication, so it might be
better done as part of a larger interface/back-end combined
refactoring.

This commit also tidies up a few other minor things in
manage_items.js, including adding automatic refresh when notes are
added/edited/deleted, and also removing a duplicate function.

Alerting the notes in the batch-receive interface is still TODO, but
should be fairly simple to add once this commit is in place.

Signed-off-by: Dan Wells <dbw2@calvin.edu>